Fix JSON parsing error and add some comments
Fix an error that was caused by projects without any programming language associated with them. For example: searching for the user "maria" would result in a parsing error because some of her projects are not associated with any language. The GithubRepository struct was fixed to address this issue. The language property is now optional.